cluster setting
